+++ src/components/task-form/confirm-step/ConfirmStep.tsx
@@ -3,7 +3,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 import s from "../TaskForm.module.scss";
 import m from "./ConfirmStep.module.scss";
 import FormsHeader from "../../ui/forms-header/FormsHeader";
-import TaskResult from "../../ui/task-result/taskResult";
+import TaskResult from "../../ui/task-result/taskResult1";
 import { Button } from "../../ui/buttons/Button";
 
 const ConfirmStep = () => {
