+++ src/components/task-form/third-step/ThirdStep.tsx
@@ -4,7 +4,7 @@ import s from "../TaskForm.module.scss";
 import m from "./ThirdStep.module.scss";
 import FormsHeader from "../../ui/forms-header/FormsHeader";
 import Select, { IOption } from "../../ui/Select/Select";
-import TextArea from "../../ui/text-area/textArea";
+import TextArea from "../../ui/text-area/textArea1";
 import { Button } from "../../ui/buttons/Button";
 
 const ThirdStep = () => {
