+++ src/components/ui/text-area/textArea1.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect, useRef, useState } from "react";
 import useAutosizeTextArea from "../../../hooks/useAutosizeTextArea";
-import s from "./textArea.module.scss";
+import s from "./textArea1.module.scss";
 
 interface TextAreaProps {
   value: string;
